move environment map rendering-related code into its own files under Renderer
